feat: split server configuration into modules
