+++ backend/package.json
@@ -7,8 +7,8 @@
     "test": "vitest",
     "build": "tsc",
     "start": "npm run build && node dist/server.js",
-    "dev": "ts-node-dev --env-file=.env --respawn --transpile-only --ignore-watch src/swagger-schemas.ts src/server.ts",
-    "dev-without-swagger": "SKIP_SWAGGER=1 ts-node-dev --env-file=.env --respawn --transpile-only --ignore-watch src/swagger-schemas.ts src/server.ts",
+    "dev": "ts-node-dev --env-file=.env --respawn --transpile-only --ignore-watch src/swagger-schemas.ts",
+    "dev-without-swagger": "SKIP_SWAGGER=1 ts-node-dev --env-file=.env --respawn --transpile-only --ignore-watch src/swagger-schemas.ts",
     "generate-erd": "ts-node src/scripts/generateErd.ts",
     "lint": "tsc --noEmit && eslint .",
     "lint:fix": "eslint . --fix",
